refactor: move article-lead-asset to its own dir, and remove unused deps
