+++ packages/article/src/article-lead-asset/get-lead-asset.js
