+++ packages/article/src/article-lead-asset/index.js
@@ -1,11 +1,11 @@
 import React from "react";
 import PropTypes from "prop-types";
-import ArticleLeadAssetVideo, {
-  propTypes as videoPropTypes
-} from "./article-lead-asset-video";
 import ArticleLeadAssetImage, {
   propTypes as imagePropTypes
 } from "./article-lead-asset-image";
+import ArticleLeadAssetVideo, {
+  propTypes as videoPropTypes
+} from "./article-lead-asset-video";
 
 const ArticleLeadAsset = ({ data }) => {
   const LeadAsset = data.isVideo
